Updated copy and print file view
Updated the text blocks that display files available to copy and print.

Files available to copy updates when the program launches, the drive changes, when the refresh drives button is clicked, and when the drive unlock button is clicked. This references data on the current drive selected.

Files available to print updates when the program launches, after files are copied, and when text is changed in the precinct text box. This references data on the local computer in the current precinct folder.