Added option for user to change file extensions copied from flash drive and made user input for this information more robust
Users can now change the types of files they are looking to copy from a flash drive in the app.  Settings are stored in C:\Temp\settings.csv
I added an error message that doesn't allow the user to use the program if the settings file is not present and provided information to explain the error.
User input handling for drive letter options and extension handling prevents user from entering in bad data and gives errors depending on what was entered.